script for dataframe and expression api docs
